Added `verbatim` parameter and list support to `zepl#send` function
The `zepl#send` function can now accept a list of strings to send to the
REPL.

The function also now has an optional parameter: `verbatim`, which
defaults to `0`.  If `verbatim` is set to `0` it will attempt to repair
and clean up malformed input by adding missing newlines and trimming
excess whitespace.  If `verbatim` is set to `1`, it will NOT modify the
input.