Add zepl-contrib
As zepl.vim aims to be both lightweight and language agnostic, it can't
really include any language-specific code, and it's feature-set is kept
to a minimum.  The downside of this is that a lot of developers (myself
included) need these additional features and better language-specific
support from their tools.

Zepl-contrib is my solution.  It allows me to ship optional features
while keeping the core "true to its goals".

As far as I know, zepl.vim is the first Vim plugin to have such
a system, so we'll see where this leads.